# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Avo Api",
    "version": "v1"
  },
  "paths": {
    "/{id}": {
      "get": {
        "tags": [
          "Avocado"
        ],
        "description": "Returns the avocado model with the corresponding id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/page": {
      "get": {
        "tags": [
          "Avocado"
        ],
        "description": "Gets a page of results",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Bags/types": {
      "get": {
        "tags": [
          "Bags"
        ],
        "description": "Returns a string array of the supported bag types",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Bags/filter/{type}": {
      "get": {
        "tags": [
          "Bags"
        ],
        "description": "Filters all days by type of bag, min number of bags sold, and an optional max number of bags sold.",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minValue",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "maxValue",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "page",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Date/year": {
      "get": {
        "tags": [
          "Date"
        ],
        "description": "returns all results in the year",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Date/after": {
      "get": {
        "tags": [
          "Date"
        ],
        "description": "returns all results after the specified date",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Date/before": {
      "get": {
        "tags": [
          "Date"
        ],
        "description": "returns all results before the specified date",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Price/range": {
      "get": {
        "tags": [
          "Price"
        ],
        "description": "returns all days where the average price is between the two values",
        "parameters": [
          {
            "name": "lowerBound",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "upperBound",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "page",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Price/gt": {
      "get": {
        "tags": [
          "Price"
        ],
        "description": "returns all the days where the average price is above the value",
        "parameters": [
          {
            "name": "value",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "page",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Price/lt": {
      "get": {
        "tags": [
          "Price"
        ],
        "description": "returns all the days where the average price is less than the value",
        "parameters": [
          {
            "name": "value",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "page",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Region/regions": {
      "get": {
        "tags": [
          "Region"
        ],
        "description": "returns all the distinct regions in the dataset",
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Region/regions/{region}": {
      "get": {
        "tags": [
          "Region"
        ],
        "description": "returns all results in the specified region",
        "parameters": [
          {
            "name": "region",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Type/types": {
      "get": {
        "tags": [
          "Type"
        ],
        "description": "Returns the supported types of avocados",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Type/types/{type}": {
      "get": {
        "tags": [
          "Type"
        ],
        "description": "returns all days with the supported type",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Volume/types": {
      "get": {
        "tags": [
          "Volume"
        ],
        "description": "returns the supported PLU types for avocado and an 'all' option to filter by all types",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/Volume/filter/{type}": {
      "get": {
        "tags": [
          "Volume"
        ],
        "description": "return all days where the volume sold is above the specified value for the PLU type. if all is passed in then it will select all days where the total volume sold is above the value",
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "vol",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "number",
              "format": "double"
            }
          },
          {
            "name": "page",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "components": { }
}